ui: move more of util files to utils subdir
