+++ ui/src/blockchain/Merkle.e2e.js
@@ -1,4 +1,4 @@
-import assert from './assert'
+import assert from '../utils/assert'
 import Merkle from './Merkle'
 
 export default {
