+++ ui/src/blockchain/Peer.js
@@ -1,5 +1,5 @@
 import Peer from 'simple-peer'
-import assert from './assert'
+import assert from '../utils/assert'
 
 // keep hardcoded STUN servers and additional TURN server in case all other P2P connection option fails
 // Peer.config.iceServers.push({urls: 'turn:numb.viagenie.ca', username: 'peer2peer@pokemail.net', credential: 'peer2peer'})
