+++ ui/src/utils/test-runner.js
@@ -1,10 +1,10 @@
 import assert from './assert.e2e'
 import crypto from './crypto.e2e'
-import Merkle from './Merkle.e2e'
-import Tx from './Tx.e2e'
-import Block from './Block.e2e'
-import Peer from './Peer.e2e'
-import PeerNetwork from './PeerNetwork.e2e'
+import Merkle from '../blockchain/Merkle.e2e'
+import Tx from '../blockchain/Tx.e2e'
+import Block from '../blockchain/Block.e2e'
+import Peer from '../blockchain/Peer.e2e'
+import PeerNetwork from '../blockchain/PeerNetwork.e2e'
 import config from '../data/config'
 import bugsnag from '../data/bugsnag'
